+++ include/Screen.hpp
@@ -5,13 +5,12 @@ class Screen
 {
 private:
 public:
-
     Vector2 GetWindowSize()
     {
         return Game::WindowSize;
     }
 
-   /**
+    /**
      * Type the percentage and it will return the Horizontal position in your screen.
      * 
      * @param Percentage is the position you want in your screen
@@ -19,7 +18,7 @@ public:
     **/
     int DynamicHPosition(float Percentage)
     {
-        return (Percentage * Game::WindowSize.X)/100;
+        return static_cast<int>((Percentage * Game::WindowSize.X) / 100.0f);
     }
 
     /**
@@ -30,6 +29,6 @@ public:
     **/
     int DynamicVPosition(float Percentage)
     {
-        return (Percentage * Game::WindowSize.Y)/100;
+        return static_cast<int>((Percentage * Game::WindowSize.Y) / 100.0f);
     }
 };
